Refactor tests by type (geometry, stiffness, tension, mass)
